save funkcja although nie dziala

+++ Dziedziczenie/Dziedziczenie/ClassPROTOCOL.cs
@@ -11,13 +11,13 @@ namespace Dziedziczenie
         string nameSurname;
         string dateBirth, dateExamination, dateCommencement;
         string chairperson, supervisor, reviewer;
-        float defenseGrade;
+        double defenseGrade;
         string[] questions;
-        float[] questionGrade;
-        float diplomaGrade;
-        float weightAverage,weightthesis,weightexam,weightx06,weightx02thesis,weightx02examx;
+        double[] questionGrade;
+        double diplomaGrade;
+        double weightAverage,weightthesis,weightexam,weightx06,weightx02thesis,weightx02examx;
         string resultWord, resultLetter;
-        float resultNumber;
+        double resultNumber;
         
         
     }
